transcribing-test/whisper-test/pyproject.toml
[project]
name = "whisper-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.43.1",
    "openai>=1.107.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
